more javadoc
