+++ src/main/java/com/teragrep/rlp_03/channel/buffer/BufferLeaseImpl.java
@@ -49,7 +49,7 @@ import java.nio.ByteBuffer;
 import java.util.concurrent.Phaser;
 
 /**
- * Implementation of the {@link BufferLease} that automatically clears (frees) the encapsulated {@link ByteBuffer} and
+ * Decorator for {@link BufferContainer} that automatically clears (frees) the encapsulated {@link ByteBuffer} and
  * returns the {@link BufferContainer} to {@link BufferLeasePool} when reference count hits zero. Starts with one
  * initial reference. Internally uses a {@link Phaser} to track reference count in a non-blocking way.
  */
