+++ src/main/java/com/teragrep/rlp_03/channel/context/EstablishedContextStub.java
@@ -45,7 +45,6 @@
  */
 package com.teragrep.rlp_03.channel.context;
 
-import com.teragrep.rlp_03.channel.InterestOps;
 import com.teragrep.rlp_03.channel.socket.Socket;
 
 import java.nio.channels.SelectionKey;
