+++ src/main/java/com/teragrep/rlp_03/channel/socket/EncryptionInfoTLS.java
@@ -43,17 +43,55 @@
  * Teragrep, the applicable Commercial License may apply to this file if you as
  * a licensee so wish it.
  */
-package com.teragrep.rlp_03.channel.info;
+package com.teragrep.rlp_03.channel.socket;
 
-public interface TransportInfo {
+import tlschannel.TlsChannel;
 
-    String getLocalAddress();
+import javax.net.ssl.SSLPeerUnverifiedException;
+import javax.security.cert.X509Certificate;
+import java.security.Principal;
+import java.security.cert.Certificate;
 
-    int getLocalPort();
+class EncryptionInfoTLS implements EncryptionInfo {
 
-    String getPeerAddress();
+    private final TlsChannel tlsChannel;
 
-    int getPeerPort();
+    public EncryptionInfoTLS(TlsChannel tlsChannel) {
+        this.tlsChannel = tlsChannel;
+    }
 
-    EncryptionInfo getEncryptionInfo();
+    @Override
+    public boolean isEncrypted() {
+        return true;
+    }
+
+    @Override
+    public String getSessionCipherSuite() {
+        return tlsChannel.getSslEngine().getSession().getCipherSuite();
+    }
+
+    @Override
+    public Certificate[] getLocalCertificates() {
+        return tlsChannel.getSslEngine().getSession().getLocalCertificates();
+    }
+
+    @Override
+    public Principal getLocalPrincipal() {
+        return tlsChannel.getSslEngine().getSession().getLocalPrincipal();
+    }
+
+    @Override
+    public X509Certificate[] getPeerCertificateChain() throws SSLPeerUnverifiedException {
+        return tlsChannel.getSslEngine().getSession().getPeerCertificateChain();
+    }
+
+    @Override
+    public Certificate[] getPeerCertificates() throws SSLPeerUnverifiedException {
+        return tlsChannel.getSslEngine().getSession().getPeerCertificates();
+    }
+
+    @Override
+    public Principal getPeerPrincipal() throws SSLPeerUnverifiedException {
+        return tlsChannel.getSslEngine().getSession().getPeerPrincipal();
+    }
 }
