+++ src/main/java/com/teragrep/rlp_03/channel/socket/Socket.java
@@ -45,8 +45,6 @@
  */
 package com.teragrep.rlp_03.channel.socket;
 
-import com.teragrep.rlp_03.channel.info.TransportInfo;
-
 import java.io.IOException;
 import java.nio.ByteBuffer;
 import java.nio.channels.SocketChannel;
