+++ src/test/java/com/teragrep/rlp_03/channel/socket/SocketFake.java
@@ -45,7 +45,6 @@
  */
 package com.teragrep.rlp_03.channel.socket;
 
-import com.teragrep.rlp_03.channel.info.TransportInfo;
 import com.teragrep.rlp_03.TransportInfoFake;
 
 import java.nio.ByteBuffer;
